[AIRFLOW-6206] Move and rename AWS batch operator [AIP-21] (#6764)
- conform to AIP-21
  - see https://issues.apache.org/jira/browse/AIRFLOW-4733
  - see https://cwiki.apache.org/confluence/display/AIRFLOW/AIP-21%3A+Changes+in+import+paths
  - use airflow.providers.amazon.aws.operators.batch.AwsBatchOperator
  - deprecate airflow.contrib.operators.awsbatch_operator.AWSBatchOperator

- fix pylint for airflow/providers/amazon/aws/operators/batch.py
